+++ core/src/tests/wasm/test_idb_storage.rs
@@ -13,7 +13,7 @@ use crate::storage::persistence::PersistenceStorageOperation;
 use crate::storage::persistence::PersistenceStorageReadAndWrite;
 use crate::storage::persistence::PersistenceStorageRemove;
 
-async fn create_db_instance(cap: usize) -> IDBStorage {
+async fn create_db_instance(cap: u32) -> IDBStorage {
     let instance = IDBStorage::new_with_cap(cap).await.unwrap();
     let (tx, store) = instance.get_tx_store(TransactionMode::ReadWrite).unwrap();
     store.clear().await.unwrap();
